Move NPM-specific code out of normalizeVcsUrl() into the NPM class
Create a new static NPM.expandShortcutURL() helper function that gets
called by NPM.parseVcsInfo() before the Package is created. This way the
Package does not need to know about the SemverType anymore.

Also move tests to the analyzer module accordingly. Remove one
PackageTest that is not applicable anymore.
